音声認識機能のリファクタリング
